Move GetPod from pod.go to taskrun.go
GetPod is only called from taskrun.go, so let's just move it to that
package and unexport it.
