+++ pkg/reconciler/taskrun/resources/pod.go
@@ -36,7 +36,6 @@ import (
 	"github.com/tektoncd/pipeline/pkg/names"
 	"github.com/tektoncd/pipeline/pkg/reconciler/taskrun/entrypoint"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -216,23 +215,6 @@ func makeWorkingDirInitializer(shellImage string, steps []v1alpha1.Step) *v1alph
 		}}
 	}
 	return nil
-}
-
-// GetPod returns the Pod for the given pod name
-type GetPod func(string, metav1.GetOptions) (*corev1.Pod, error)
-
-// TryGetPod fetches the TaskRun's pod, returning nil if it has not been created or it does not exist.
-func TryGetPod(taskRunStatus v1alpha1.TaskRunStatus, gp GetPod) (*corev1.Pod, error) {
-	if taskRunStatus.PodName == "" {
-		return nil, nil
-	}
-
-	pod, err := gp(taskRunStatus.PodName, metav1.GetOptions{})
-	if err == nil || errors.IsNotFound(err) {
-		return pod, nil
-	}
-
-	return nil, err
 }
 
 // MakePod converts TaskRun and TaskSpec objects to a Pod which implements the taskrun specified
