+++ pkg/reconciler/taskrun/resources/pod_test.go
@@ -28,10 +28,8 @@ import (
 	"github.com/tektoncd/pipeline/test/names"
 	"golang.org/x/xerrors"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	fakek8s "k8s.io/client-go/kubernetes/fake"
 )
 
@@ -42,68 +40,6 @@ var (
 	credsImage = "override-with-creds:latest"
 	shellImage = "busybox"
 )
-
-func TestTryGetPod(t *testing.T) {
-	err := xerrors.New("something went wrong")
-	for _, c := range []struct {
-		desc    string
-		trs     v1alpha1.TaskRunStatus
-		gp      GetPod
-		wantNil bool
-		wantErr error
-	}{{
-		desc: "no-pod",
-		trs:  v1alpha1.TaskRunStatus{},
-		gp: func(string, metav1.GetOptions) (*corev1.Pod, error) {
-			t.Errorf("Did not expect pod to be fetched")
-			return nil, nil
-		},
-		wantNil: true,
-		wantErr: nil,
-	}, {
-		desc: "non-existent-pod",
-		trs: v1alpha1.TaskRunStatus{
-			PodName: "no-longer-exist",
-		},
-		gp: func(name string, opts metav1.GetOptions) (*corev1.Pod, error) {
-			return nil, errors.NewNotFound(schema.GroupResource{}, name)
-		},
-		wantNil: true,
-		wantErr: nil,
-	}, {
-		desc: "existing-pod",
-		trs: v1alpha1.TaskRunStatus{
-			PodName: "exists",
-		},
-		gp: func(name string, opts metav1.GetOptions) (*corev1.Pod, error) {
-			return &corev1.Pod{}, nil
-		},
-		wantNil: false,
-		wantErr: nil,
-	}, {
-		desc: "pod-fetch-error",
-		trs: v1alpha1.TaskRunStatus{
-			PodName: "something-went-wrong",
-		},
-		gp: func(name string, opts metav1.GetOptions) (*corev1.Pod, error) {
-			return nil, err
-		},
-		wantNil: true,
-		wantErr: err,
-	}} {
-		t.Run(c.desc, func(t *testing.T) {
-			pod, err := TryGetPod(c.trs, c.gp)
-			if err != c.wantErr {
-				t.Fatalf("TryGetPod: %v", err)
-			}
-
-			wasNil := pod == nil
-			if wasNil != c.wantNil {
-				t.Errorf("Pod got %v, want %v", wasNil, c.wantNil)
-			}
-		})
-	}
-}
 
 func TestMakePod(t *testing.T) {
 	names.TestingSeed()
